+++ src/lib.rs
@@ -126,7 +126,16 @@ pub fn trace(configs: &[Config]) -> Result<TraceMap, RunError> {
 
 fn create_logger(configs: &[Config]) -> Option<EventLog> {
     if configs.iter().any(|c| c.dump_traces) {
-        Some(EventLog::new(configs.iter().map(|x| x.root()).collect()))
+        let config = if let Some(c) = configs.iter().find(|c| c.output_directory.is_some()) {
+            c
+        } else {
+            &configs[0]
+        };
+
+        Some(EventLog::new(
+            configs.iter().map(|x| x.root()).collect(),
+            config,
+        ))
     } else {
         None
     }
@@ -229,18 +238,14 @@ pub fn launch_tarpaulin(
     if !config.no_run {
         let project_analysis = SourceAnalysis::get_analysis(config);
         let project_analysis = project_analysis.lines;
+        let mut other_bins = config.objects().to_vec();
+        other_bins.extend(executables.binaries.iter().cloned());
         for exe in &executables.test_binaries {
             if exe.should_panic() {
                 info!("Running a test executable that is expected to panic");
             }
-            let coverage = get_test_coverage(
-                exe,
-                &executables.binaries,
-                &project_analysis,
-                config,
-                false,
-                logger,
-            )?;
+            let coverage =
+                get_test_coverage(exe, &other_bins, &project_analysis, config, false, logger)?;
             if let Some(res) = coverage {
                 result.merge(&res.0);
                 return_code |= if exe.should_panic() {
@@ -250,14 +255,8 @@ pub fn launch_tarpaulin(
                 };
             }
             if config.run_ignored {
-                let coverage = get_test_coverage(
-                    exe,
-                    &executables.binaries,
-                    &project_analysis,
-                    config,
-                    true,
-                    logger,
-                )?;
+                let coverage =
+                    get_test_coverage(exe, &other_bins, &project_analysis, config, true, logger)?;
                 if let Some(res) = coverage {
                     result.merge(&res.0);
                     return_code |= res.1;